requirements.txt
Flask==3.0.*
flask-cors==4.0.*
mysql-connector-python==9.1.*
bcrypt==4.1.*
gunicorn==22.0.0